backend/requirements.txt
fastapi
pypdf
google-generativeai
python-multipart
uvicorn
python-dotenv
psycopg2-binary
numpy
pydantic
